# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AgingOut": {
        "properties": {
          "as_of": {
            "format": "date",
            "title": "As Of",
            "type": "string"
          },
          "bucket_order": {
            "items": {
              "$ref": "#/components/schemas/BucketLabel"
            },
            "title": "Bucket Order",
            "type": "array"
          },
          "company_timezone": {
            "title": "Company Timezone",
            "type": "string"
          },
          "no_bill_details": {
            "items": {
              "$ref": "#/components/schemas/NoBillDetails"
            },
            "title": "No Bill Details",
            "type": "array"
          },
          "parties": {
            "items": {
              "$ref": "#/components/schemas/AgingRow"
            },
            "title": "Parties",
            "type": "array"
          },
          "side": {
            "enum": [
              "receivable",
              "payable"
            ],
            "title": "Side",
            "type": "string"
          },
          "total": {
            "$ref": "#/components/schemas/AgingRow"
          },
          "unverified_gates": {
            "items": {
              "type": "string"
            },
            "title": "Unverified Gates",
            "type": "array"
          }
        },
        "required": [
          "side",
          "as_of",
          "company_timezone",
          "bucket_order",
          "total",
          "parties",
          "no_bill_details",
          "unverified_gates"
        ],
        "title": "AgingOut",
        "type": "object"
      },
...
      "AllocationsOut": {
        "properties": {
          "allocations": {
            "items": {
              "$ref": "#/components/schemas/AllocationOut"
            },
            "title": "Allocations",
            "type": "array"
          },
          "as_of": {
            "format": "date",
            "title": "As Of",
            "type": "string"
          },
          "ledger_id": {
            "format": "uuid",
            "title": "Ledger Id",
            "type": "string"
          },
          "reference_name": {
            "title": "Reference Name",
            "type": "string"
          },
          "side": {
            "enum": [
              "receivable",
              "payable"
            ],
            "title": "Side",
            "type": "string"
          },
          "unverified_gates": {
            "items": {
              "type": "string"
            },
            "title": "Unverified Gates",
            "type": "array"
          }
        },
        "required": [
          "side",
          "as_of",
          "ledger_id",
          "reference_name",
          "allocations",
          "unverified_gates"
        ],
        "title": "AllocationsOut",
        "type": "object"
      },
...
      "BillsOut": {
        "properties": {
          "as_of": {
            "format": "date",
            "title": "As Of",
            "type": "string"
          },
          "bills": {
            "items": {
              "$ref": "#/components/schemas/app__schemas__aging__BillOut"
            },
            "title": "Bills",
            "type": "array"
          },
          "ledger_id": {
            "format": "uuid",
            "title": "Ledger Id",
            "type": "string"
          },
          "side": {
            "enum": [
              "receivable",
              "payable"
            ],
            "title": "Side",
            "type": "string"
          },
          "unverified_gates": {
            "items": {
              "type": "string"
            },
            "title": "Unverified Gates",
            "type": "array"
          }
        },
        "required": [
          "side",
          "as_of",
          "ledger_id",
          "bills",
          "unverified_gates"
        ],
        "title": "BillsOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MetricName": {
        "enum": [
          "sales",
          "customer-revenue",
          "product-revenue",
          "product-difference",
          "purchases",
          "supplier-purchases",
          "expenses",
          "cash-flow",
          "cash-bank-position",
          "receivables",
          "payables",
          "ledger-balances",
          "unclassified-adjustments"
        ],
        "title": "MetricName",
        "type": "string"
      },
      "MetricOut": {
        "properties": {
          "breakdown": {
            "items": {
              "$ref": "#/components/schemas/BreakdownRow"
            },
            "title": "Breakdown",
            "type": "array"
          },
          "company_timezone": {
            "title": "Company Timezone",
            "type": "string"
          },
          "filters_applied": {
            "$ref": "#/components/schemas/FiltersApplied"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/SeriesPoint"
            },
            "title": "Series",
            "type": "array"
          },
          "summary": {
            "$ref": "#/components/schemas/Figure"
          }
        },
        "required": [
          "metric",
          "filters_applied",
          "company_timezone",
          "summary",
          "series",
          "breakdown",
          "notes"
        ],
        "title": "MetricOut",
        "type": "object"
      },
...
      "PaymentBehaviourOut": {
        "description": "FR-PAY-1-6, D-049 #5. Until gate G25 passes, `available` is false and nothing is\ncomputed (FR-PAY-6).",
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "customers": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PaymentFigureOut"
            },
            "title": "Customers",
            "type": "array"
          },
          "excluded_settlements": {
            "additionalProperties": {
              "type": "integer"
            },
            "default": {},
            "title": "Excluded Settlements",
            "type": "object"
          },
          "min_settlements": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Settlements"
          },
          "notes": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "overall": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaymentFigureOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "unverified_gates": {
            "items": {
              "type": "string"
            },
            "title": "Unverified Gates",
            "type": "array"
          },
          "window_from": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window From"
          },
          "window_to": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window To"
          }
        },
        "required": [
          "available",
          "unverified_gates"
        ],
        "title": "PaymentBehaviourOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Tally Analytics Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/companies/{company_id}/analytics/aging": {
      "get": {
        "operationId": "aging_summary_companies__company_id__analytics_aging_get",
        "parameters": [
          {
            "in": "path",
            "name": "company_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Company Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "side",
            "required": true,
            "schema": {
              "enum": [
                "receivable",
                "payable"
              ],
              "title": "Side",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Receivables or payables aged per bill (SRS 10)",
        "tags": [
          "analytics"
        ]
      }
    },
    "/companies/{company_id}/analytics/aging/allocations": {
      "get": {
        "operationId": "aging_allocations_companies__company_id__analytics_aging_allocations_get",
        "parameters": [
          {
            "in": "path",
            "name": "company_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Company Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "side",
            "required": true,
            "schema": {
              "enum": [
                "receivable",
                "payable"
              ],
              "title": "Side",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "ledger_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ledger Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "reference",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Reference",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllocationsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "A bill's allocations and vouchers",
        "tags": [
          "analytics"
        ]
      }
    },
    "/companies/{company_id}/analytics/aging/bills": {
      "get": {
        "operationId": "aging_bills_companies__company_id__analytics_aging_bills_get",
        "parameters": [
          {
            "in": "path",
            "name": "company_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Company Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "side",
            "required": true,
            "schema": {
              "enum": [
                "receivable",
                "payable"
              ],
              "title": "Side",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "ledger_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ledger Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "A party's bills (aging drill-down)",
        "tags": [
          "analytics"
        ]
      }
    },
...
    "/companies/{company_id}/analytics/payment-behaviour": {
      "get": {
        "operationId": "payment_behaviour_companies__company_id__analytics_payment_behaviour_get",
        "parameters": [
          {
            "in": "path",
            "name": "company_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Company Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentBehaviourOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Customer payment behaviour (SRS 10.4)",
        "tags": [
          "analytics"
        ]
      }
    },
...
    "/companies/{company_id}/analytics/stock": {
      "get": {
        "operationId": "stock_movement_companies__company_id__analytics_stock_get",
        "parameters": [
          {
            "in": "path",
            "name": "company_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Company Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "period_days",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Period Days"
            }
          },
          {
            "in": "query",
            "name": "class",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "FAST",
                    "NORMAL",
                    "SLOW",
                    "DEAD",
                    "NEVER_SOLD",
                    "STOCK_UNKNOWN",
                    "NOT_CLASSIFIED"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Class"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Stock movement classes (SRS 11)",
        "tags": [
          "analytics"
        ]
      }
    },
...
    "/companies/{company_id}/analytics/{metric}": {
      "get": {
        "operationId": "metric_companies__company_id__analytics__metric__get",
        "parameters": [
          {
            "in": "path",
            "name": "metric",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/MetricName"
            }
          },
          {
            "in": "path",
            "name": "company_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Company Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "granularity",
            "required": false,
            "schema": {
              "default": "month",
              "enum": [
                "day",
                "month",
                "quarter"
              ],
              "title": "Granularity",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "group_by",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Group By"
            }
          },
          {
            "in": "query",
            "name": "include_cancelled",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Cancelled",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_missing",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Missing",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "customer",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Customer"
            }
          },
          {
            "in": "query",
            "name": "product",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product"
            }
          },
          {
            "in": "query",
            "name": "cost_centre",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cost Centre"
            }
          },
          {
            "description": "Repeatable '<group_by option>:<key>'; key 'none' is the NULL bucket",
            "in": "query",
            "name": "by",
            "required": false,
            "schema": {
              "default": [],
              "description": "Repeatable '<group_by option>:<key>'; key 'none' is the NULL bucket",
              "items": {
                "type": "string"
              },
              "title": "By",
              "type": "array"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetricOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Metric",
        "tags": [
          "analytics"
        ]
      }
    },
    "/companies/{company_id}/analytics/{metric}/drilldown": {
      "get": {
        "operationId": "drilldown_companies__company_id__analytics__metric__drilldown_get",
        "parameters": [
          {
            "in": "path",
            "name": "metric",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/MetricName"
            }
          },
          {
            "in": "path",
            "name": "company_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Company Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "include_cancelled",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Cancelled",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_missing",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Missing",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "customer",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Customer"
            }
          },
          {
            "in": "query",
            "name": "product",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product"
            }
          },
          {
            "in": "query",
            "name": "cost_centre",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cost Centre"
            }
          },
          {
            "description": "Repeatable '<group_by option>:<key>'; key 'none' is the NULL bucket",
            "in": "query",
            "name": "by",
            "required": false,
            "schema": {
              "default": [],
              "description": "Repeatable '<group_by option>:<key>'; key 'none' is the NULL bucket",
              "items": {
                "type": "string"
              },
              "title": "By",
              "type": "array"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DrilldownOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Drilldown",
        "tags": [
          "analytics"
        ]
      }
    },
...
  }
}